# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Algerian Supermarket Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all employees (Admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create employee (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update employee (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete employee (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - name: cashierId
          in: query
          schema:
            type: integer
        - name: customerId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sales list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create sale (Cashier)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleBody"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get sale by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
...
  /shortages:
    get:
      operationId: listShortages
      tags: [shortages]
      summary: List shortage/damage reports
      responses:
        "200":
          description: Shortage reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShortageReport"
    post:
      operationId: createShortage
      tags: [shortages]
      summary: Report shortage or damage (Worker)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShortageBody"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortageReport"
  /shortages/{id}/resolve:
    patch:
      operationId: resolveShortage
      tags: [shortages]
      summary: Resolve shortage report (Admin/Buyer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveShortageBody"
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortageReport"
  /salaries:
    get:
      operationId: listSalaries
      tags: [salaries]
      summary: List salary records (Admin only)
      parameters:
        - name: month
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Salary records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalaryRecord"
    post:
      operationId: createSalaryRecord
      tags: [salaries]
      summary: Create/pay salary (Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSalaryBody"
      responses:
        "201":
          description: Salary record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalaryRecord"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, cashier, buyer, worker]
        baseSalary:
          type: ["number", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [cashier, buyer, worker]
        baseSalary:
          type: ["number", "null"]
        phone:
          type: ["string", "null"]
      required:
        - name
        - email
        - password
        - role
    UpdateUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        baseSalary:
          type: ["number", "null"]
        phone:
          type: ["string", "null"]
...
    SaleItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        quantity:
          type: number
        unit:
          type: string
        subtotal:
          type: number
      required:
        - productId
        - productName
        - price
        - quantity
        - unit
        - subtotal
    Sale:
      type: object
      properties:
        id:
          type: integer
        cashierId:
          type: integer
        cashierName:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        paid:
          type: boolean
        paymentMethod:
          type: string
          enum: [cash, karni, card]
        createdAt:
          type: string
      required:
        - id
        - cashierId
        - cashierName
        - items
        - subtotal
        - discount
        - total
        - paid
        - paymentMethod
        - createdAt
    CreateSaleBody:
      type: object
      properties:
        cashierId:
          type: integer
        customerId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: number
              unit:
                type: string
            required:
              - productId
              - quantity
              - unit
        discount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, karni, card]
      required:
        - cashierId
        - items
        - paymentMethod
...
    ShortageReport:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: ["integer", "null"]
        productName:
          type: string
        reportedById:
          type: integer
        reportedByName:
          type: string
        type:
          type: string
          enum: [shortage, damage, expired]
        quantity:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, resolved, dismissed]
        resolvedById:
          type: ["integer", "null"]
        resolvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - productName
        - reportedById
        - reportedByName
        - type
        - status
        - createdAt
    CreateShortageBody:
      type: object
      properties:
        productId:
          type: ["integer", "null"]
        productName:
          type: string
        type:
          type: string
          enum: [shortage, damage, expired]
        quantity:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - productName
        - type
    ResolveShortageBody:
      type: object
      properties:
        status:
          type: string
          enum: [resolved, dismissed]
        notes:
          type: ["string", "null"]
      required:
        - status
    SalaryRecord:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userRole:
          type: string
        month:
          type: string
        baseSalary:
          type: number
        bonus:
          type: number
        deduction:
          type: number
        netSalary:
          type: number
        paid:
          type: boolean
        paidAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - userName
        - userRole
        - month
        - baseSalary
        - bonus
        - deduction
        - netSalary
        - paid
        - createdAt
    CreateSalaryBody:
      type: object
      properties:
        userId:
          type: integer
        month:
          type: string
        baseSalary:
          type: number
        bonus:
          type: number
        deduction:
          type: number
        paid:
          type: boolean
        notes:
          type: ["string", "null"]
      required:
        - userId
        - month
        - baseSalary
...